cronet import: set gclient --delete_unversioned_trees
This option deletes gclient dependencies that were remove upstream, so
it makes sense to always set this flag for the import (since the import
directoy is reused to speed up the process).

This change also fixes setup_folder_origin to run in a subshell to
contain the directory change it performs.

Test: none
Change-Id: I87c60690be81843c3947acc0251967372ed4c34a